--- flask/app/templates/customer.html
@@ -1,20 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<div class="container">
-  <div class="row">
-    <h2>Add Customer to Our Database</h2>
-    <form action="" method="post" name="login">
-      {{ form.hidden_tag() }}
-      <p>
-        Please enter the company name:<br>
-        {{ form.company(size=120) }}<br>
-      </p>
-      <p>
-        Please enter the user email:<br>
-        {{ form.email(size=120) }}<br>
-      </p>
-      <p><input type="submit" value="Create Customer"></p>
-    </form>
-  </div>
-</div>
-{% endblock %}
