+++ flask/orm-tutorial/models-tutorial/app/forms.py
@@ -1,0 +1,9 @@
+from flask.ext.wtf import Form
+from wtforms import StringField, IntegerField
+from flask_wtf.html5 import EmailField
+from wtforms.validators import DataRequired
+
+class CustomerForm(Form):
+    company = StringField('company', validators=[DataRequired()])
+    email = EmailField('email', validators=[DataRequired()])
+    # Add additional Address fields here
