+++ flask/orm-tutorial/models-tutorial/app/templates/home.html
@@ -17,6 +17,8 @@
             </tr>
           </thead>
           {% for customer in customers %}
+          <!-- {% for address in customer.addresses %}
+          {% endfor %} -->
             <tr>
               <td>{{ customer.company }}</td>
               <td>{{ customer.email }}</td>
