+++ flask/orm-tutorial/models-tutorial/config.py
